ENH: Removed logic mixin, since most methods are not useful for mpReview
