+++ Util/mixins.py
@@ -150,49 +150,3 @@ class ModuleWidgetMixin(object):
         logging.error("qMRMLNodeComboBox does not have attribute %s" % key)
     combobox.setMRMLScene(slicer.mrmlScene)
     return combobox
-
-
-class ModuleLogicMixin(object):
-
-  @staticmethod
-  def createDirectory(directory, message=None):
-    if message:
-      logging.debug(message)
-    try:
-      os.makedirs(directory)
-    except OSError:
-      logging.debug('Failed to create the following directory: ' + directory)
-
-  @staticmethod
-  def getDICOMValue(currentFile, tag, fallback=None):
-    db = slicer.dicomDatabase
-    try:
-      value = db.fileValue(currentFile, tag)
-    except RuntimeError:
-      logging.info("There are problems with accessing DICOM values from file %s" % currentFile)
-      value = fallback
-    return value
-
-  @staticmethod
-  def getFileList(directory):
-    return [f for f in os.listdir(directory) if ".DS_Store" not in f]
-
-  @staticmethod
-  def importStudy(dicomDataDir):
-    indexer = ctk.ctkDICOMIndexer()
-    indexer.addDirectory(slicer.dicomDatabase, dicomDataDir)
-    indexer.waitForImportFinished()
-
-  @staticmethod
-  def createScalarVolumeNode(name):
-    volume = slicer.vtkMRMLScalarVolumeNode()
-    volume.SetName(name)
-    slicer.mrmlScene.AddNode(volume)
-    return volume
-
-  @staticmethod
-  def createTransformNode(name, isBSpline):
-    node = slicer.vtkMRMLBSplineTransformNode() if isBSpline else slicer.vtkMRMLLinearTransformNode()
-    node.SetName(name)
-    slicer.mrmlScene.AddNode(node)
-    return node
